modif: added subnav to manufac and products

+++ shopyo/manage.py
@@ -62,14 +62,24 @@ def runserver():
 
 @manager.command
 def rundebug():
-    app.run()
+    app.run(debug=True, host='0.0.0.0')
 
-@manager.command
 def clean():
-    os.remove('test.db')
-    shutil.rmtree('__pycache__')
-    shutil.rmtree('migrations')
-
+    if os.path.exists('test.db'):
+        os.remove('test.db')
+        print("test.db successfully deleted")
+    else:   
+        print("test.db doesn't exist")
+    if os.path.exists('__pycache__'):
+        shutil.rmtree('__pycache__')
+        print("__pycache__ successfully deleted")
+    else:
+        print("__pycache__ doesn't exist")
+    if os.path.exists('migrations'):
+        shutil.rmtree('migrations')
+        print("migrations successfully deleted")
+    else:
+        print("migrations folder doesn't exist")
 
 if __name__ == '__main__':
     manager.run()
